Uploading early Python Project
This is an early Python project I did. It is a file syncing program
that was written with versions of Python, Paramiko, and Mac OS from a
couple years ago so it is very unlikely to be able to run now.

I added a README describing the project and included screenshots and a
screen recording from when it was first completed.